requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
akshare==1.17.4
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
decorator==5.2.1
et_xmlfile==2.0.0
frozenlist==1.7.0
html5lib==1.1
idna==3.10
jsonpath==0.82.2
lxml==5.4.0
mini-racer==0.12.4
multidict==6.4.4
nest-asyncio==1.6.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1747544610608/work/dist/numpy-2.2.6-cp310-cp310-macosx_10_13_x86_64.whl#sha256=11cc21c326fa6386ceb48e396557886bf55226b93cb330d1e1eab3a6a9968215
openpyxl==3.1.5
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1749099742855/work
propcache==0.3.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML==6.0.2
requests==2.32.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
soupsieve==2.7
ta-lib @ file:///Users/runner/miniforge3/conda-bld/ta-lib_1731944887113/work
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3==2.4.0
webencodings==0.5.1
xlrd==2.0.2
yarl==1.20.1
